pyproject.toml
[project]
name = "tsproj-ltsfm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chronos-forecasting>=2.2.2",
    "datasets>=4.8.4",
    "einops>=0.8.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "patool>=4.0.2",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "timesfm>=1.3.0",
    "tirex>=0.0.2",
    "torch>=2.10.0",
    "transformers>=4.41,<5",
]
